Extract "dynamic" and ORM tests in their own file too
This is similar to previous commits and moves out more tests from the
common tests file.

This commit is best viewed using Git --color-moved option.
